adding classes for stocks + market
